fix: 풀리퀘스트 리뷰 내용 반영
